sam2-test/requirements.txt
antlr4-python3-runtime==4.9.3
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
hydra-core==1.3.2
iopath==0.1.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
portalocker==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pywin32==308
PyYAML==6.0.2
sam2==0.4.1
setuptools==75.5.0
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
